frontend/requirements.txt
flask
python-dotenv
flask-login
flask_wtf
email_validator
requests
PyJWT==2.1.0
